requirements.txt
python-dotenv~=1.0.1
aiohttp~=3.10.10
aiogram~=3.13.1